% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_networks.R
\name{check_config_spatialNW}
\alias{check_config_spatialNW}
\title{check_config_spatialNW}
\usage{
check_config_spatialNW(config, spat_locs)
}
\arguments{
\item{config}{a list of config would be used to create spatial network using delaunay method only via Giotto::createSpatialNetwork}

\item{spat_locs}{a data.frame with spatial location info}
}
\value{
the corrected config list
}
\description{
check config used to create spatial network, assign default values if missing arguments
}
