% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{checkFileExists}
\alias{checkFileExists}
\title{checkFileExists}
\usage{
checkFileExists(Annotation, name)
}
\arguments{
\item{Annotation}{the data.frame storing the sample annotation file}

\item{name}{the column name in Annotation}
}
\value{
a message about which file is missing
}
\description{
function to check if file is missing in the Annotation data.frame
}
