% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline.R
\name{choose_distance_cutoff}
\alias{choose_distance_cutoff}
\title{choose_distance_cutoff}
\usage{
choose_distance_cutoff(
  transcript_df,
  transID_coln = "transcript_id",
  cellID_coln = "cell_ID",
  spatLocs_colns = c("x", "y", "z"),
  extracellular_cellID = NULL,
  sampleSize_nROI = 10,
  sampleSize_cellNum = 2500,
  seed = 123,
  run_molecularDist = TRUE
)
}
\arguments{
\item{transcript_df}{the data.frame for each transcript}

\item{transID_coln}{the column name of transcript_ID in \code{transcript_df}}

\item{cellID_coln}{the column name of cell_ID in \code{transcript_df}}

\item{spatLocs_colns}{column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in \code{transcript_df}}

\item{extracellular_cellID}{a vector of cell_ID for extracellular transcripts which would be removed from the resegmention pipeline (default = NULL)}

\item{sampleSize_nROI}{number of ROIs randomly picked from data for molecular distance cutoff estimation}

\item{sampleSize_cellNum}{maximum number of cells from the picked ROIs for molecular distance cutoff estimation}

\item{seed}{a random seed for sub-sampling cells from whole dataset for molecular distance cutoff estimation}

\item{run_molecularDist}{flag to run molecular distant cutoff estimation, default = TRUE}
}
\value{
a list
\enumerate{
\item{cellular_distance_cutoff, maximum cell-to-cell distance in x, y between the center of query cells to the center of neighbor cells with direct contact, same unit as input spatial coordinate. }
\item{perCell_coordDT, a data.table with cell in row, spatial XY coordiantes of centroid and dimensions of bounding box in column}
\item{molecular_distance_cutoff, maximum molecule-to-molecule distance within connected transcript group, same unit as input spatial coordinate; return if run_molecularDist = TRUE}
\item{distance_profile, a named vector for the quantile profile of minimal molecular distance between transcripts belong to different cells at step size of 10\% quantile; return if run_molecularDist = TRUE}
}
}
\description{
Chose appropriate cellular distance cutoff and molecular distance cutoff based on input transcript data.frame for downstream resegmentation; cellular distance cutoff is defined as the search radius of direct neighbor cell, while molecular distance cutoff is defined as the maximum distance between two neighbor transcripts from same source cells.
}
\details{
\code{cellular_distance_cutoff} is defined as maximum cell-to-cell distance in x, y between the center of query cells to the center of neighbor cells with direct contact. The function calculates average 2D cell diameter from input data.frame and use 2 times of the mean cell diameter as \code{cellular_distance_cutoff}. \code{molecular_distance_cutoff} is defined as maximum molecule-to-molecule distance within connected transcript groups belonging to same source cells. The function would first randomly choose \code{sampleSize_cellNum} number of cells from \code{sampleSize_nROI} number of randomly picked ROIs with search radius to be 5 times of \code{cellular_distance_cutoff}, and then calculate the minimal molecular distance between picked cells. The function would further use the 10 times of 90\% quantile of minimal molecular distance as \code{molecular_distance_cutoff}. This calculation is slow and is not recommended for large transcript data.frame.
}
\examples{
data(transcriptDF)
extracellular_cellID <- transcriptDF[which(transcriptDF$CellId ==0), 'cell_ID'] # cell_ID for extracellualr transcripts
distCutoffs <- choose_distance_cutoff(transcriptDF,
                                      extracellular_cellID = extracellular_cellID)
}
