% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastReseg_full_pipeline.R
\name{compartment_and_add_extra}
\alias{compartment_and_add_extra}
\title{compartment_and_add_extra}
\usage{
compartment_and_add_extra(
  updated_transDF,
  transcript_df,
  updatedCellID_coln = "updated_cellID",
  compartment_coln = "transComp",
  combine_extra = FALSE
)
}
\arguments{
\item{updated_transDF}{transcript data.frame with transcript id and updated cell segmentation, one of the outputs  of \code{fastReseg_perFOV_full_process}}

\item{transcript_df}{list of 2 elements, \code{intraC} and \code{extraC} for intracellular and extracellular transcript data.frame, respectively; columns contain transcript id and original cell segmentation, output of \code{prepare_perFOV_transDF}}

\item{updatedCellID_coln}{column name of updated cell ID in \code{updated_transDF}, default = "updated_cellID"}

\item{compartment_coln}{column name for transcript compartment label, default = "transComp"}

\item{combine_extra}{flag to combine original extracellular transcripts and trimmed transcripts back to the updated transcript data.frame, slow process if many transcript in each FOV file. (default = FALSE)}
}
\value{
a transcript data.frame containing all original transcripts with updated cell segmentation and transcript compartments for \code{intraC}, \code{extraC} and \code{trimmed}
}
\description{
add compartment label to the updated transcript data.frame returned by \code{fastReseg_perFOV_full_process} and optional to combine it with transcripts got trimmed and originally extracellular transcripts
}
