% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_networks.R
\name{createSpatialDelaunayNW_from_spatLocs}
\alias{createSpatialDelaunayNW_from_spatLocs}
\title{createSpatialDelaunayNW_from_spatLocs}
\usage{
createSpatialDelaunayNW_from_spatLocs(
  config_spatNW,
  spatLocs_df,
  ID_column = "cell_ID",
  spatLocs_column = c("sdimx", "sdimy", "sdimz")
)
}
\arguments{
\item{config_spatNW}{configuration list}

\item{spatLocs_df}{data.frame for spatial location of each entry for cell or transcript}

\item{ID_column}{column name for entry ID in spatLocs_df}

\item{spatLocs_column}{column name for 1st, 2nd, optional 3rd dimension of spatial coordinates in spatLocs_df}
}
\value{
delaunay_network_Obj, a spatial network object created by Giotto function
}
\description{
generate delaunay network based on provided config and spatial location using Giotto functions
}
