% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_transcripts.R
\name{estimate_MeanProfile}
\alias{estimate_MeanProfile}
\title{estimate_MeanProfile}
\usage{
estimate_MeanProfile(counts, clust, s, bg)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{clust}{Vector of cluster assignments, or a matrix of probabilities of cells (rows) belonging to clusters (columns).}

\item{s}{Vector of scaling factors for each cell, e.g. as defined by cell area.}

\item{bg}{Expected background}
}
\value{
A matrix of cluster profiles, genes * clusters
}
\description{
Given cell assignments (or posterior probabilities), estimate the mean profile of each cluster.
}
\details{
E step: estimate each cluster's mean profile (Ptolemy Estep function)

estimate each cluster's mean profile (Ptolemy Estep function)
}
