% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_refProfiles}
\alias{example_refProfiles}
\title{Data for cell type specific reference profiles}
\format{
An object of class \code{"matrix"} with 960 genes in rows, 22 distinct cell types in columns.
}
\usage{
data(example_refProfiles)
}
\description{
Semi-supervised cell typing is conducted on the example SMI RNA readout for FFPE melanoma tissue samples that contains 2 tissue sections with 3 FOVs per section. The resulting cell typing outcomes are then used to estimate the mean profile of each identified cell type/cluster using \code{"estimate_MeanProfile"}. The corresponding average cell-type-specific profiles are then stored in \code{data("example_refProfiles")} to be used as reference profiles.
}
\examples{
data(example_refProfiles)
}
\keyword{datasets}
