% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastReseg_perFOV_full_process.R
\name{fastReseg_perFOV_full_process}
\alias{fastReseg_perFOV_full_process}
\title{fastReseg_perFOV_full_process}
\usage{
fastReseg_perFOV_full_process(
  score_GeneMatrix,
  transcript_df,
  transID_coln = "UMI_transID",
  transGene_coln = "target",
  cellID_coln = "UMI_cellID",
  spatLocs_colns = c("x", "y", "z"),
  extracellular_cellID = NULL,
  flagModel_TransNum_cutoff = 50,
  flagCell_lrtest_cutoff = 5,
  svmClass_score_cutoff = -2,
  svm_args = list(kernel = "radial", scale = FALSE, gamma = 0.4),
  molecular_distance_cutoff = 2.7,
  cellular_distance_cutoff = NULL,
  score_baseline = NULL,
  lowerCutoff_transNum = NULL,
  higherCutoff_transNum = NULL,
  groupTranscripts_method = c("dbscan", "delaunay"),
  spatialMergeCheck_method = c("leidenCut", "geometryDiff"),
  cutoff_spatialMerge = 0.5,
  leiden_config = list(objective_function = "CPM", resolution_parameter = 1, beta =
    0.01, n_iterations = 200),
  config_spatNW_transcript = NULL,
  return_intermediates = TRUE,
  return_perCellData = TRUE,
  includeAllRefGenes = FALSE,
  seed_process = NULL
)
}
\arguments{
\item{score_GeneMatrix}{the gene x cell-type matrix of log-like score of gene in each cell type}

\item{transcript_df}{the data.frame for each transcript with columns for transcript_id, target or gene name, original cell_id, spatial coordinates.}

\item{transID_coln}{the column name of transcript_ID in \code{transcript_df}}

\item{transGene_coln}{the column name of target or gene name in \code{transcript_df}}

\item{cellID_coln}{the column name of cell_ID in \code{transcript_df}}

\item{spatLocs_colns}{column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in \code{transcript_df}}

\item{extracellular_cellID}{a vector of cell_ID for extracellular transcripts which would be removed from the resegmention pipeline (default = NULL)}

\item{flagModel_TransNum_cutoff}{the cutoff of transcript number to do spatial modeling for identification of wrongly segmented cells (default = 50)}

\item{flagCell_lrtest_cutoff}{the cutoff of lrtest_nlog10P to identify putative wrongly segemented cells with strong spatial dependency in transcript score profile}

\item{svmClass_score_cutoff}{the cutoff of transcript score to separate between high and low score transcripts in SVM (default = -2)}

\item{svm_args}{a list of arguments to pass to svm function for identifying low-score transcript groups in space, typically involve kernel, gamma, scale}

\item{molecular_distance_cutoff}{maximum molecule-to-molecule distance within connected transcript group, same unit as input spatial coordinate (default = 2.7 micron).
If set to NULL, the pipeline would first randomly choose no more than 2500 cells from up to 10 random picked ROIs with search radius to be 5 times of \code{cellular_distance_cutoff}, and then calculate the minimal molecular distance between picked cells. The pipeline would further use the 5 times of 90\% quantile of minimal molecular distance as \code{molecular_distance_cutoff}. This calculation is slow and is not recommended for large transcript data.frame.}

\item{cellular_distance_cutoff}{maximum cell-to-cell distance in x, y between the center of query cells to the center of neighbor cells with direct contact, same unit as input spatial coordinate. Default = NULL to use the 2 times of average 2D cell diameter.}

\item{score_baseline}{a named vector of score baseline under each cell type listed in \code{score_GeneMatrix} such that per cell transcript score higher than the baseline is required to call a cell type of high enough confidence}

\item{lowerCutoff_transNum}{a named vector of transcript number cutoff under each cell type such that higher than the cutoff is required to keep query cell as it is}

\item{higherCutoff_transNum}{a named vector of transcript number cutoff under each cell type such that lower than the cutoff is required to keep query cell as it is when there is neighbor cell of consistent cell type.}

\item{groupTranscripts_method}{use either "dbscan" or "delaunay method" to group transcripts in space (default = "dbscan")}

\item{spatialMergeCheck_method}{use either "leidenCut" (in 2D or 3D) or "geometryDiff" (in 2D only) method to determine whether a cell pair merging event is allowed in space (default = "leidenCut")}

\item{cutoff_spatialMerge}{spatial constraint on a valid merging event between two source transcript groups, default = 0.5 for 50\% cutoff, set to 0 to skip spatial constraint evaluation for merging.
For \code{spatialMergeCheck_method = "leidenCut"}, this is the minimal percentage of transcripts shared membership between query cell and neighbor cells in leiden clustering results for a valid merging event.
For \code{spatialMergeCheck_method = "geometryDiff"}, this is the maximum percentage of white space change upon merging of query cell and neighbor cell for a valid merging event.}

\item{leiden_config}{(leidenCut) a list of configuration to pass to reticulate and \code{igraph::cluster_leiden} function, including objective_function, resolution_parameter, beta, n_iterations.}

\item{config_spatNW_transcript}{configuration list to create spatial network at transcript level, see manual for \code{createSpatialDelaunayNW_from_spatLocs} for more details, set to NULL to use default config (default = NULL)}

\item{return_intermediates}{flag to return intermediate outputs, including data.frame for spatial modeling statistics of each cell}

\item{return_perCellData}{flag to return gene x cell count matrix and per cell DF with updated mean spatial coordinates and new cell type}

\item{includeAllRefGenes}{flag to include all genes in \code{score_GeneMatrix} in the returned \code{updated_perCellExprs} with missing genes of value 0 (default = FALSE)}

\item{seed_process}{seed for per FOV processing, used in transcript error detection and correction steps, default = NULL to skip the seed}
}
\value{
a list
\describe{
\item{modStats_ToFlagCells}{a data.frame for spatial modeling statistics of each cell, output of \code{score_cell_segmentation_error} function, return when \code{return_intermediates} = TRUE}
\item{groupDF_ToFlagTrans}{data.frame for the group assignment of transcripts within putative wrongly segmented cells, merged output of \code{flag_bad_transcripts} and \code{groupTranscripts_Delaunay} or \code{groupTranscripts_dbscan} functions, return when \code{return_intermediates} = TRUE}
\item{neighborhoodDF_ToReseg}{a data.frame for neighborhood environment of low-score transcript groups, output of \code{get_neighborhood_content} function, return when \code{return_intermediates} = TRUE}
\item{reseg_actions}{a list of 4 elements describing how the resegmenation would be performed on original \code{transcript_df} by the group assignment of transcripts listed in \code{groupDF_ToFlagTrans}, output of \code{decide_ReSegment_Operations} function, return when \code{return_intermediates} = TRUE}
\item{updated_transDF}{the updated transcript_df with \code{updated_cellID} and \code{updated_celltype} column based on reseg_full_converter}
\item{updated_perCellDT}{a per cell data.table with mean spatial coordinates, new cell type and resegmentation action after resegmentation, return when \code{return_perCellData} = TRUE}
\item{updated_perCellExprs}{a gene x cell count sparse matrix for updated transcript data.frame after resegmentation, return when \code{return_perCellData} = TRUE}
}
}
\description{
core wrapper for resegmentation pipeline using transcript score matrix derived from external reference profiles and preset cutoffs
}
\details{
The pipeline would score each transcript based on the provided cell type-specific reference profiles, evaluate the goodness-of-fit of each transcript within original cell segment, identify the low-score transcript groups within cells that has strong spatial dependency in transcript score profile, evaluate the neighborhood environment of low-score transcript groups and perform resegmentation actions including triming to extracellular space, merging to neighbor cell or labeling as new cell.

To account for genes missing in \code{score_GeneMatrix} but present in input transcript data.frame, genes in \code{ctrl_genes} would be assigned with goodness-of-fit score equal to \code{svmClass_score_cutoff} for all cell types to minimize the impact of those genes on the identification of low-score transcript groups via SVM. To avoid significant interference from those \code{ctrl_genes}, it's recommended to have total counts of those genes below 1\% of total counts of all genes in each cell.
}
\examples{
data(example_refProfiles)
data(mini_transcriptDF)
data(example_baselineCT)
extracellular_cellID <- mini_transcriptDF[which(mini_transcriptDF$CellId ==0), 'cell_ID'] #' cell_ID for extracellualr transcripts
score_baseline <- example_baselineCT[['span_score']][,"25\%"]
lowerCutoff_transNum  <- example_baselineCT[['span_transNum']][,"25\%"]
higherCutoff_transNum  <- example_baselineCT[['span_transNum']][,"50\%"]
#' calculate log-likelihood of each gene under each cell type and center the score matrix on per gene basis
score_GeneMatrix <- scoreGenesInRef(genes = intersect(unique(mini_transcriptDF[["target"]]), rownames(example_refProfiles)),
                                    ref_profiles = pmax(example_refProfiles, 1e-5))
# case 1: run with default methods: "dbscan" for transcript grouping, "leidenCut" for merging check 
final_res1 <- fastReseg_perFOV_full_process(score_GeneMatrix= score_GeneMatrix,
                                           transcript_df = mini_transcriptDF,
                                           extracellular_cellID = extracellular_cellID,
                                           molecular_distance_cutoff = 2.7,
                                           cellular_distance_cutoff = 25,
                                           score_baseline = score_baseline,
                                           lowerCutoff_transNum = lowerCutoff_transNum,
                                           higherCutoff_transNum= higherCutoff_transNum)


# case 2: run with alternative methods: "delaunay" for transcript grouping, "geometryDiff" for merging check 
final_res2 <- fastReseg_perFOV_full_process(score_GeneMatrix= score_GeneMatrix,
                                           transcript_df = mini_transcriptDF,
                                           extracellular_cellID = extracellular_cellID,
                                           molecular_distance_cutoff = 2.7,
                                           cellular_distance_cutoff = 25,
                                           score_baseline = score_baseline,
                                           lowerCutoff_transNum = lowerCutoff_transNum,
                                           higherCutoff_transNum= higherCutoff_transNum,
                                           groupTranscripts_method = "delaunay",
                                           spatialMergeCheck_method = "geometryDiff")


# case 3: run with default "dbscan" for transcript grouping, but apply no spatial constraint on merging
final_res3 <- fastReseg_perFOV_full_process(score_GeneMatrix= score_GeneMatrix,
                                            transcript_df = mini_transcriptDF,
                                            extracellular_cellID = extracellular_cellID,
                                            molecular_distance_cutoff = 2.7,
                                            cellular_distance_cutoff = 25,
                                            score_baseline = score_baseline,
                                            lowerCutoff_transNum = lowerCutoff_transNum,
                                            higherCutoff_transNum= higherCutoff_transNum,
                                            cutoff_spatialMerge = 0)
}
