% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_errors.R
\name{flagTranscripts_LDA_hyperplane}
\alias{flagTranscripts_LDA_hyperplane}
\title{flagTranscripts_LDA_hyperplane}
\usage{
flagTranscripts_LDA_hyperplane(
  chosen_cells,
  score_GeneMatrix,
  transcript_df,
  cellID_coln = "CellId",
  transID_coln = "transcript_id",
  transGene_coln = "target",
  score_coln = "score",
  spatLocs_colns = c("x", "y", "z"),
  model_cutoff = 50,
  score_cutoff = -2
)
}
\arguments{
\item{chosen_cells}{the cell_ID of chosen cells}

\item{score_GeneMatrix}{the gene x cell-type matrix of log-like score of gene in each cell type}

\item{transcript_df}{the data.frame of transcript_ID, cell_ID, score, spatial coordinates}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transGene_coln}{the column name of target or gene name in transcript_df}

\item{score_coln}{the column name of score in transcript_df}

\item{model_cutoff}{the cutoff of transcript number to do spatial modeling (default = 50)}

\item{score_cutoff}{the cutoff of score to separate between high and low score transcripts (default = -2)}

\item{spatLocs_coln}{the column name of 1st, 2nd, optional 3rd spatial dimension of each transcript in transcript_df}
}
\value{
a data.frame
\enumerate{
\item{cellID_coln, original cell id}
\item{LDA_class, 0 for below cutoff, 1 for above cutoff}
\item{transID_coln, original transcript_id}
\item{transGene_coln, target gene of transcript}
\item{score_coln, score in transcript_df}
\item{spatLocs_colns for spatial coorindates of transcript}
\item{LD1, LD1 value of LDA model output}
\item{LDA_cell_type, new cell type for each transcript groups within each cells}
}
}
\description{
find out the spatially connected transcripts among chosen_transcripts based on LDA hyperplane spatial model which scores each cell for how much their transcripts change their goodness-of-fit over space.
}
\details{
For score of transcripts within each cell, assign 0 or 1 label to each transcript based on whether the score is above score_cutoff; then run linear discriminant analysis on lda(above_cutoff ~ x + y + xy + x^2 + y^2) for 2D, lda(above_cutoff ~ x + y + z+ xy +xz + yz + x^2 + y^2 + z^2) for 3D. Coordinate variables with variance less than 1e-8 would not be used. Z-step of 5E-4mm or similar often resulted in too small of variables for z2.
}
