% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_score_utilities.R
\name{getCellType_maxScore}
\alias{getCellType_maxScore}
\title{getCellType_maxScore}
\usage{
getCellType_maxScore(
  score_GeneMatrix,
  transcript_df,
  transGene_coln = "target",
  cellID_coln = "cell_ID"
)
}
\arguments{
\item{score_GeneMatrix}{a gene x cell-type score matrix}

\item{transcript_df}{the data.frame of transcript_ID and cell_ID}

\item{transGene_coln}{the column name of target or gene name in transcript_df}

\item{cellID_coln}{the column name of cell_ID in transcript_df}
}
\value{
return a named vector with cell type in values and cell_ID in names
}
\description{
get the cell type give maximum score
}
