% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{getNeighbors_transDF}
\alias{getNeighbors_transDF}
\title{getNeighbors_transDF}
\usage{
getNeighbors_transDF(
  chosen_cells = NULL,
  cell_networkDT = NULL,
  network_cellID_split = "_g",
  neighbor_distance_xy = NULL,
  neighbor_distance_z = NULL,
  transcript_df,
  cellID_coln = "CellId",
  transID_coln = "transcript_id",
  transSpatLocs_coln = c("x", "y", "z"),
  gridSpat_coln = c("fov", "slide")
)
}
\arguments{
\item{chosen_cells}{the cell_ID of chosen cells need to be evaluate for re-segmentation}

\item{cell_networkDT}{the data.table component of cell-level delaunay network object, would be used for neighborhood search if neighbor_distance_xy is NULL.}

\item{network_cellID_split}{the deliminator used to convert cellID_coln in transcript_df to the cell ID in cell_networkDT}

\item{neighbor_distance_xy}{the distance in x, y from the outermost transcripts of each chosen cell to the furthest non-self transcripts for neighborhood network building at transcript level. Default = NULL to use the cell network for selection of neighborhood.}

\item{neighbor_distance_z}{the distance in z direction from the outermost transcripts of each chosen cell to the furthest non-self transcripts for neighborhood network building at transcript level. Default = NULL to use the cell network for selection of neighborhood.}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transSpatLocs_coln}{the column name of 1st, 2nd, optional 3rd spatial dimension of each transcript in transcript_df}

\item{gridSpat_coln}{an optional vector of column names in transcript_df that could act as grid to separate different transcripts in space, such as columns for FOV, slideID when using direct distance threshold to define neighborhood search range.}
}
\value{
a data.frame
\enumerate{
\item{transcript_id, transcript id of all transcript near chosen cell's neighborhood}
\item{transSpatLocs_coln, coordinates}
\item{in_query_cell, flag whether current transcript is in the query cell, false means in neighbor cells}
\item{CellId, original cell id of each transcript}
\item{query_CellId, original query cell id of transcript's neighborhood}
}
}
\description{
find neighbor cells of chosen_cells and return the relevant transcript data.frame for both query
}
\details{
get neighborhood transcript data frame for visualization

If no neighbor cells found for query cell, return query cell information only. Do not consider extracellular transcripts.
}
