% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_transcripts.R
\name{getScoreCellType_gene}
\alias{getScoreCellType_gene}
\title{getScoreCellType_gene}
\usage{
getScoreCellType_gene(
  score_GeneMatrix,
  transcript_df,
  transID_coln = "transcript_id",
  transGene_coln = "target",
  celltype_coln = "cell_type"
)
}
\arguments{
\item{score_GeneMatrix}{a gene x cell-type score matrix}

\item{transcript_df}{the data.frame of transcript_ID and cell_ID}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transGene_coln}{the column name of target or gene name in transcript_df}

\item{celltype_coln}{the column name of cell type in transcript_df}
}
\value{
score_df, a data.frame of "\link{transID_coln}" and "score_\link{celltype_coln}" column for chosen cell-type
}
\description{
get each transcript's score based on score matrix and chosen cell-type
}
