% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_transcripts_in_space.R
\name{groupTranscripts_dbscan}
\alias{groupTranscripts_dbscan}
\title{groupTranscripts_dbscan}
\usage{
groupTranscripts_dbscan(
  chosen_transcripts = NULL,
  distance_cutoff = "auto",
  transcript_df,
  cellID_coln = "CellId",
  transID_coln = "transcript_id",
  transSpatLocs_coln = c("x", "y", "z")
)
}
\arguments{
\item{chosen_transcripts}{the transcript_id of chosen transcript}

\item{distance_cutoff}{maximum molecule-to-molecule distance within same transcript group (default = "auto")}

\item{transcript_df}{the data.frame with transcript_id, target/geneName, x, y and cell_id}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transSpatLocs_coln}{the column name of 1st, 2nd, optional 3rd spatial dimension of each transcript in transcript_df}
}
\value{
data frame of connected transcripts among chosen_transcripts
#' \enumerate{
\item{cellID_coln, orignal cell_ID}
\item{transID_coln, connected transcripts among chosen_transcripts}
\item{transSpatLocs_coln, spatial coordiantes of transcript}
\item{transcript_group, group of chosen_transcripts}
}
}
\description{
group the flagged transcript within each cell based on spatial clustering using dbscan
}
\details{
For query cell, group flagged transcripts only based on their molecular distance to each other. When distance cutoff = 'auto', use 20\% average XY cell range as cutoff. In case of no more than 3 flagged transcripts per cell, determine the grouping based on distance cutoff directly. In case of more transcripts per cell, use \code{dbscan} to group transcripts with distance_cutoff as \code{eps} and \code{minPts = 1}.
}
