% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_errors.R
\name{groupTranscripts_Delaunay}
\alias{groupTranscripts_Delaunay}
\title{groupTranscripts_Delaunay}
\usage{
groupTranscripts_Delaunay(
  chosen_transcripts = NULL,
  config_spatNW_transcript = list(name = "transcript_delaunay_network", dimensions =
    "all", method = "Delaunay", minimum_k = 0, delaunay_method = "delaunayn_geometry",
    maximum_distance_delaunay = "auto", options = "Pp", Y = TRUE, j = TRUE, S = 0),
  distance_cutoff = "auto",
  transcript_df,
  cellID_coln = "CellId",
  transID_coln = "transcript_id",
  transSpatLocs_coln = c("x", "y", "z"),
  visualize_network = FALSE
)
}
\arguments{
\item{chosen_transcripts}{the transcript_id of chosen transcript}

\item{config_spatNW_transcript}{configuration list to create spatial network at transcript level, see manual for \code{createSpatialDelaunayNW_from_spatLocs} for more details}

\item{distance_cutoff}{maximum distance within connected transcript group (default = "auto")}

\item{transcript_df}{the data.frame with transcript_id, target/geneName, x, y and cell_id}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transSpatLocs_coln}{the column name of 1st, 2nd, optional 3rd spatial dimension of each transcript in transcript_df}
}
\value{
data frame of connected transcripts among chosen_transcripts
#' \enumerate{
\item{cellID_coln, orignal cell_ID}
\item{transID_coln, connected transcripts among chosen_transcripts}
\item{transSpatLocs_coln, spatial coordiantes of transcript}
\item{transcript_group, group of chosen_transcripts}
}
}
\description{
group the flagged transcript within each cell based on spatial connectivity of their transcript delaunay network
}
\details{
for query cell, build network on flagged transcripts only to identify groups. In case of no more than 3 transcripts, determine the grouping based on distance cutoff directly; when distance cutoff = 'auto', no additional edge filtering based on delaunay network output but use 20\% average XY cell range as cutoff when no more than 3 transcript.
}
