% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_transcripts_in_space.R
\name{igraph_delete_edges}
\alias{igraph_delete_edges}
\title{igraph_delete_edges}
\usage{
igraph_delete_edges(graph, edges)
}
\arguments{
\item{graph}{An igraph object from which edges will be deleted.}

\item{edges}{A vector of edge IDs or an edge selector to delete.}
}
\value{
An igraph object with the specified edges removed.
}
\description{
Delete edges from an igraph object with version compatibility
}
\details{
This function wraps \code{igraph::delete_edges()} and ensures compatibility
with older versions (<2.0.0) of \code{igraph} that used \code{delete.edges()}.
}
