% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_celltyping.R
\name{lldist}
\alias{lldist}
\title{lldist}
\usage{
lldist(x, mat, bg = 0.01, size = 10, digits = 2)
}
\arguments{
\item{x}{a vector of a reference cell cluster}

\item{mat}{a cells x genes matrix of expression levels in all cells}

\item{bg}{a vector for background level of each cell (default: 0.01)}

\item{size}{the parameters for dnbinom function (default: 10)}

\item{digits}{the number of digits for rounding}
}
\value{
a cell x cell_type matrix of the log-likelihood
}
\description{
calculate the log-likelihood of cell belong to certain cell cluster given the reference profile using negative binomial model
}
