% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastReseg_perFOV_full_process.R
\name{makeDummyOuts_perFOV}
\alias{makeDummyOuts_perFOV}
\title{makeDummyOuts_perFOV}
\usage{
makeDummyOuts_perFOV(
  all_genes,
  transcript_df,
  transID_coln = "UMI_transID",
  transGene_coln = "target",
  cellID_coln = "UMI_cellID",
  spatLocs_colns = c("x", "y", "z"),
  return_intermediates = TRUE,
  return_perCellData = TRUE,
  includeAllRefGenes = FALSE
)
}
\arguments{
\item{all_genes}{vector of all genes consider in the analysis as listed in rows of \code{score_GeneMatrix} in parent function}

\item{transcript_df}{the data.frame for each transcript with columns for transcript_id, target or gene name, original cell_id, spatial coordinates, as well as \code{tLLR_maxCellType} and \code{score_tLLR_maxCellType} columns from earlier step in parent function}

\item{transID_coln}{the column name of transcript_ID in \code{transcript_df}}

\item{transGene_coln}{the column name of target or gene name in \code{transcript_df}}

\item{cellID_coln}{the column name of cell_ID in \code{transcript_df}}

\item{spatLocs_colns}{column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in \code{transcript_df}}

\item{return_intermediates}{flag to return intermediate outputs, including data.frame for spatial modeling statistics of each cell}

\item{return_perCellData}{flag to return gene x cell count matrix and per cell DF with updated mean spatial coordinates and new cell type}

\item{includeAllRefGenes}{flag to include all genes in \code{score_GeneMatrix} in the returned \code{updated_perCellExprs} with missing genes of value 0 (default = FALSE)}
}
\value{
a list
\describe{
\item{groupDF_ToFlagTrans}{data.frame for the group assignment of transcripts within putative wrongly segmented cells, merged output of \code{flag_bad_transcripts} and \code{groupTranscripts_Delaunay} or \code{groupTranscripts_dbscan} functions, return when \code{return_intermediates} = TRUE}
\item{neighborhoodDF_ToReseg}{a data.frame for neighborhood environment of low-score transcript groups, output of \code{get_neighborhood_content} function, return when \code{return_intermediates} = TRUE}
\item{reseg_actions}{a list of 4 elements describing how the resegmenation would be performed on original \code{transcript_df} by the group assignment of transcripts listed in \code{groupDF_ToFlagTrans}, output of \code{decide_ReSegment_Operations} function, return when \code{return_intermediates} = TRUE}
\item{updated_transDF}{the updated transcript_df with \code{updated_cellID} and \code{updated_celltype} column based on reseg_full_converter}
\item{updated_perCellDT}{a per cell data.table with mean spatial coordinates, new cell type and resegmentation action after resegmentation, return when \code{return_perCellData} = TRUE}
\item{updated_perCellExprs}{a gene x cell count sparse matrix for updated transcript data.frame after resegmentation, return when \code{return_perCellData} = TRUE}
}
}
\description{
supporting function for \code{fastReseg_perFOV_full_process}, makes dummy outputs based on input \code{transcript_df} in same format as the outputs of \code{fastReseg_perFOV_full_process()} but returns no \code{modStats_ToFlagCells} data.frame, used in case of no cells being flagged for cell segmentation error.
}
