% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_transcriptDF}
\alias{mini_transcriptDF}
\title{Transcript level data frame from Example dataset for spatial transcriptional profiling of tissue}
\format{
An object of class \code{"data.frame"} with 756783 transcripts in row and 9 variables
\describe{
\item{transcript_id}{unique id for transcript}
\item{cell_ID}{unique id for cell based on original cell segmentaion assignment}
\item{x}{spatial coordinate of the transcript in x-axis of the given FOV, unit in micron}
\item{y}{spatial coordinate of the transcript in y-axis of the given FOV, unit in micron}
\item{z}{spatial coordinate of the transcript in z-axis of the given FOV, unit in micron}
\item{target}{gene identity of given transcript}
\item{slide}{slide ID of given transcript}
\item{fov}{fov ID of given transcript}
\item{CellId}{the cell label assignment of given transcript within given FOV based on original cell segmentation}
}
}
\usage{
data(mini_transcriptDF)
}
\description{
The example dataset is based on one run of SMI RNA readout for FFPE melanoma tissue samples. It contains information for 1375 cells with 756783 transcripts within a cropped region of one single FOV of 1 tissue section.
}
\keyword{datasets}
