% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_transcripts_in_space.R
\name{myFun_3point_singleCell}
\alias{myFun_3point_singleCell}
\title{myFun_3point_singleCell}
\usage{
myFun_3point_singleCell(
  dfCoord_subset,
  transSpatLocs_coln = c("x", "y", "z"),
  distance_cutoff = 2.7,
  startGroup = 1
)
}
\arguments{
\item{dfCoord_subset}{transcript data.table for single cell with only 3 transcripts in rows}

\item{transSpatLocs_coln}{the column name of 1st, 2nd, optional 3rd spatial dimension of each transcript in transcript_df}

\item{distance_cutoff}{maximum distance within connected transcript group}

\item{startGroup}{the index of starting group ID}
}
\value{
a data.table with \code{transcript_group} column added to original input data.table
}
\description{
supporting function for \code{groupTranscripts_Delaunay}, assign group ID for 3 transcripts in single cell in 3D based on distant cutoff
}
