% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{myFun_fov_load}
\alias{myFun_fov_load}
\title{myFun_fov_load}
\usage{
myFun_fov_load(path_to_fov)
}
\description{
supporting function for \code{runPreprocess}, \code{fastReseg_internalRef} and \code{fastReseg_flag_all_errors}, to load transcript data.frame of each FOV from file path
}
