% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{numCores}
\alias{numCores}
\title{Get number of cores for parallelized operations}
\usage{
numCores(percentCores = 0.9, minNotUsedCores = 2)
}
\arguments{
\item{percentCores}{percent of cores to use for parallelization, value range 0 to 1}

\item{minNotUsedCores}{minimum number of cores to leave for background processes}
}
\value{
number of cores to use for mclapply
}
\description{
Get number of cores for parallelized operations
}
