% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ori_CellStatsDF}
\alias{ori_CellStatsDF}
\title{Cell level data frame from Example dataset for spatial transcriptional profiling of tissue}
\format{
An object of class \code{"data.frame"} with 1375 cells in row and 10 variables
\describe{
\item{cell_ID}{unique id for cell based on original cell segmentaion assignment}
\item{slide}{slide ID of given cell}
\item{fov}{fov ID of given cell}
\item{CellId}{the cell label assignment of given cell within given FOV based on original cell segmentation}
\item{CenterX}{spatial coordinate of the cell centroid in x-axis of the given FOV, unit in micron}
\item{CenterY}{spatial coordinate of the cell centroid in y-axis of the given FOV, unit in micron}
\item{Width}{the bounding box length of given cell segment along x-axis, unit in micron}
\item{Height}{the bounding box length of given cell segment along y-axis, unit in micron}
\item{Area}{the area of given cell segment, unit in square micron}
\item{AspectRatio}{the aspect ratio of boundingbox width over height for given cell segment}
}
}
\usage{
data(ori_CellStatsDF)
}
\description{
The example dataset is based on one run of SMI RNA readout for FFPE melanoma tissue samples. It contains information for 1375 cells with 756783 transcripts within a cropped region of one single FOV of 1 tissue section.
}
\keyword{datasets}
