% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ori_RawExprs}
\alias{ori_RawExprs}
\title{Cell x Gene expression matrix from Example dataset for spatial transcriptional profiling of tissue}
\format{
An object of class \code{"matrix"} with 4619 cells in row and 960 genes in columns
}
\usage{
data(ori_RawExprs)
}
\description{
The example dataset is based on one run of SMI RNA readout for FFPE melanoma tissue samples. It contains the single cell gene expression profile for 1375 cells across 960 genes.
}
\keyword{datasets}
