% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_auxiliary.R
\name{plotSpatialScoreMultiCells}
\alias{plotSpatialScoreMultiCells}
\title{plotSpatialScoreMultiCells}
\usage{
plotSpatialScoreMultiCells(
  chosen_cells,
  cell_labels,
  transcript_df,
  cellID_coln = "CellId",
  transID_coln = "transcript_id",
  score_coln = "score",
  spatLocs_colns = c("x", "y")
)
}
\arguments{
\item{chosen_cells}{the cell_ID of chosen cells}

\item{cell_labels}{title labels for each cell's spatial plot}

\item{transcript_df}{the data.frame of transcript_ID, cell_ID, spatial locations}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{score_coln}{the column name of score in transcript_df}

\item{spatLocs_colns}{column names for 1st, 2nd dimension of spatial coordinates in transcript_df}
}
\value{
ggplot
}
\description{
plot spatial plot of transcripts within chosen cells, colored by score, facet by cell_ID
}
