% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepare_perFOV_transDF}
\alias{prepare_perFOV_transDF}
\title{prepare_perFOV_transDF}
\usage{
prepare_perFOV_transDF(
  each_transDF,
  fov_centerLocs,
  prefix_vals = NULL,
  pixel_size = 0.18,
  zstep_size = 0.8,
  transID_coln = NULL,
  transGene_coln = "target",
  cellID_coln = "CellId",
  spatLocs_colns = c("x", "y", "z"),
  extracellular_cellID = NULL,
  drop_original = FALSE
)
}
\arguments{
\item{each_transDF}{data.frame for raw transcript}

\item{fov_centerLocs}{a named vector of fov 2D coordinates}

\item{prefix_vals}{a named vector of values to be used as prefix in \code{UMI_transID} and \code{UMI_cellID}; when \code{prefix_vals} != NULL, unique transcript_id would be generated from \code{prefix_vals} and \code{transID_coln} in \code{each_transDF}}

\item{pixel_size}{the micrometer size of image pixel listed in 1st and 2nd dimension of \code{spatLocs_colns} of \code{each_transDF}}

\item{zstep_size}{the micrometer size of z-step for the optional 3rd dimension of \code{spatLocs_colns} of \code{each_transDF}}

\item{transID_coln}{the column name of transcript_ID in \code{transcript_df}, default = NULL to use row index of transcript in \code{each_transDF}; when \code{prefix_vals} != NULL, unique transcript_id would be generated from \code{prefix_vals} and \code{transID_coln} in \code{each_transDF}}

\item{transGene_coln}{the column name of target or gene name in \code{each_transDF}}

\item{cellID_coln}{the column name of cell_ID in \code{each_transDF}; when \code{prefix_colns} != NULL, unique cell_ID would be generated from \code{prefix_vals} and \code{cellID_coln} in each \code{transcript_df}}

\item{spatLocs_colns}{column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in \code{each_transDF}}

\item{extracellular_cellID}{a vector of cell_ID for extracellular transcripts which would be removed from the resegmention pipeline (default = NULL)}

\item{drop_original}{flag to drop original per FOV based cell ID and coordinates under columns \code{CellId}, \code{pixel_x}, \code{pixel_y}, \code{idx_z} (default = FALSE)}
}
\value{
a list contains transcript_df for downstream process and extracellular transcript data.frame
' \describe{
\item{intraC}{a data.frame for intracellular transcript, \code{UMI_transID} and \code{UMI_cellID} as column names for unique transcript_id and cell_id, \code{target} as column name for target gene name}
\item{extraC}{a data.frame for extracellular transcript, same structure as the \code{intraC} data.frame in returned list}
}
}
\description{
supporting function for \code{runPreprocess}, \code{fastReseg_internalRef} and \code{fastReseg_flag_all_errors} to get unique IDs for cells and transcripts, and convert pixel coordinates to um; when \verb{drop_original = FALSE, the function will also return original per FOV based cell ID and coordinates under columns }CellId\verb{, }pixel_x\verb{, }pixel_y\verb{, }idx_z`.
}
