% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastReseg_perFOV_full_process.R
\name{prepResegDF}
\alias{prepResegDF}
\title{prepResegDF}
\usage{
prepResegDF(
  transcript_df,
  groupDF_ToFlagTrans,
  cellID_coln = "UMI_cellID",
  transID_coln = "UMI_transID"
)
}
\arguments{
\item{transcript_df}{the data.frame for each transcript with columns for transcript_id, target or gene name, original cell_id, spatial coordinates and \code{tLLR_maxCellType}.}

\item{groupDF_ToFlagTrans}{data frame for transcripts in \code{chosen_cells} only, with columns for \code{connect_group},\code{tmp_cellID},\code{group_maxCellType}, output of \code{runTranscriptErrorDetection()}}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}
}
\value{
a list of two elements
\describe{
\item{reseg_transcript_df}{data.frame with transcript_id, target/geneName, x, y, cell_id for all transcript groups \code{tmp_cellID} and the cell type of maximum transcript scores for each transcript group \code{group_maxCellType}}
\item{groups_to_reseg}{vector of chosen transcript groups need to be evaluate for re-segmentation}
}
}
\description{
supporting function for \code{fastReseg_perFOV_full_process}, combine \code{runTranscriptErrorDetection()} output with transcript data.frame to prep for \code{runSegRefinement()}
}
