% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepSMI_for_fastReseg}
\alias{prepSMI_for_fastReseg}
\title{prepSMI_for_fastReseg}
\usage{
prepSMI_for_fastReseg(
  path_to_SMIobject,
  config_loading,
  refClus_coln = "nb_clus",
  cellClus_to_exclude = NULL,
  removeUnpaired = FALSE,
  blacklist_genes = NULL,
  pixel_size = 0.18
)
}
\arguments{
\item{path_to_SMIobject}{file path to \code{"Giotto"} object for existing SMI multi-FOV multi-slide data that was save in \code{.RData} format}

\item{config_loading}{a list holding arguments passed to \code{SMITAP::data_loading_from_config} when creating SMI object; this list should contain file path information for the raw data used for creating the corresponding multi-slide multi-FOV object.}

\item{refClus_coln}{the column name of cell cluster assignment saved in \code{cell_metadata} of the existing SMI object}

\item{cellClus_to_exclude}{a vector of cell cluster name to be excluded for calculation of reference profiles (default = NULL)}

\item{removeUnpaired}{flag to remove FOVs with unpaired target call files and fov position information; default = FALSE, to stop processing when missing target call files}

\item{blacklist_genes}{a vector of genes to be excluded from reference profile estimation (default = NULL)}

\item{pixel_size}{the micrometer size of image pixel listed in \code{Width} and \code{Height} dimension of each cell stored in \code{cell_metadata} of the existing SMI object (default = 0.18)}
}
\value{
a list
\describe{
\item{counts}{a cells X genes count matrix for entire dataset, stored in SMI object.}
\item{clust}{vector of cluster assignments for each cell in \code{counts}, stored in SMI object.}
\item{refProfiles}{a genes X clusters matrix of cluster-specific reference profiles estimated from entire dataset.}
\item{score_baseline}{a named vector of score baseline under each cell type listed in \code{refProfiles} such that  per cell transcript score higher than the baseline is required to call a cell type of high enough confidence.}
\item{lowerCutoff_transNum}{a named vector of transcript number cutoff under each cell type such that higher than the cutoff is required to keep query cell as it is.}
\item{higherCutoff_transNum}{a named vector of transcript number cutoff under each cell type such that lower than the cutoff is required to keep query cell as it is when there is neighbor cell of consistent cell type.}
\item{cellular_distance_cutoff}{maximum cell-to-cell distance in x, y between the center of query cells to the center of neighbor cells with direct contact, unit in micron. Use the 2 times of average 2D cell diameter.}
\item{transDF_fov_fileInfo}{a data.frame with each row for each individual file of per FOV transcript data.frame within which the coordinates and CellId are unique, columns include the file path of per FOV transcript data.frame file, annotation columns like slide and fov to be used as prefix when creating unique cell_ID across entire dataset, in colnames of \code{file_path}, \code{slide},\code{slideName},\code{fov},\code{offset_x}, \code{offset_y}.}
\item{sample_annot}{a data.frame of sample annotation with each row for each slide, columns for file path and metedata that are read from file path of \code{config_loading[["annotfile"]]}.}
\item{ctrl_genes}{a vector of ctrl genes in SMI object, include all feat_ID not under \code{rna} feat_type. }
}
}
\description{
prepare resegmentation inputs from existing SMI object with multi-FOV multi-slide data
}
\details{
The function requires the pre-built \code{"Giotto"} object that contains single cell typing results on original cell segmentation outcomes. This pre-built object may contain multi-slide multi-FOV data and should be built with raw data whose file path information is stored in \code{config_loading}.
\code{config_loading} should contain at least the following 6 elements:
\describe{
\item{annotfile}{file path to the sample annotation file}
\item{folderpathColumn}{Column name for full file path to run folder}
\item{slidefoldersColumn}{Column name for base file path to slide folder in each run folder}
\item{slidenameColumn}{Column name for slide_name}
\item{votedfoldersColumn}{Column name for base file path to voted folders in each slide folder}
\item{versionColumn}{Column name for the version of the target calling pipeline}
}
The function would first calculate reference profiles, \code{refProfiles}, and several cutoffs from the entire dataset, including the cutoffs for transcript number, \code{lowerCutoff_transNum} and \code{higherCutoff_transNum}, the cutoff for transcript tLLR score \code{score_baseline}, and the center-to-center distance cutoff for neighbor cells with direct contact, \code{cellular_distance_cutoff}.
Then, the function would check the existence of fov offset position and transcript data.frame for each FOV in each slide.
Lastly, the function would prepare \code{transDF_fov_fileInfo} to be used with \code{fastReseg_internalRef} function for downstream resegmentation workflow which would use either NULL or a user-defined value for \code{molecular_distance_cutoff}, the molecule-to-molecule distance cutoff for neighbor molecules belonging to same cell. When \code{molecular_distance_cutoff} is set to NULL in downstream \code{fastReseg_internalRef} function, \code{molecular_distance_cutoff} would then be estimated from transcript data for the first FOV transcript data.frame.
}
