% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_celltyping.R
\name{quick_celltype}
\alias{quick_celltype}
\title{quick_celltype}
\usage{
quick_celltype(
  x,
  bg = 0.01,
  reference_profiles,
  nb_size = 10,
  align_genes = TRUE
)
}
\arguments{
\item{x}{Counts matrix (or dgCMatrix), cells x genes.}

\item{bg}{a vector for background level of each cell (default = 0.01)}

\item{reference_profiles}{Matrix of expression profiles of pre-defined clusters, genes x clusters.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{align_genes}{Logical, for whether to align the counts matrix and the reference_profiles by gene ID.}
}
\value{
A list, with the following elements:
\enumerate{
\item clust: a vector given cells' cluster assignments
\item logliks: Matrix of cells' log-likelihoods under each cluster. Cells in rows, clusters in columns.
}
}
\description{
Classify cells based on reference profiles given the maximum log-likelihood calculated via negative binomial model
}
