% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTranscriptErrorDetection.R
\name{runTranscriptErrorDetection}
\alias{runTranscriptErrorDetection}
\title{runTranscriptErrorDetection}
\usage{
runTranscriptErrorDetection(
  chosen_cells,
  score_GeneMatrix,
  transcript_df,
  cellID_coln = "CellId",
  transID_coln = "transcript_id",
  transGene_coln = "target",
  score_coln = "score",
  spatLocs_colns = c("x", "y", "z"),
  model_cutoff = 50,
  score_cutoff = -2,
  svm_args = list(kernel = "radial", scale = FALSE, gamma = 0.4),
  groupTranscripts_method = c("dbscan", "delaunay"),
  distance_cutoff = "auto",
  config_spatNW_transcript = NULL,
  seed_transError = NULL
)
}
\arguments{
\item{chosen_cells}{the cell_ID of chosen cells}

\item{score_GeneMatrix}{the gene x cell-type matrix of log-like score of gene in each cell type}

\item{transcript_df}{the data.frame of transcript_ID, cell_ID, score, spatial coordinates}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transGene_coln}{the column name of target or gene name in transcript_df}

\item{score_coln}{the column name of score in transcript_df}

\item{model_cutoff}{the cutoff of transcript number to do spatial modeling (default = 50)}

\item{score_cutoff}{the cutoff of score to separate between high and low score transcripts (default = -2)}

\item{svm_args}{a list of arguments to pass to svm function, typically involve kernel, gamma, scale}

\item{groupTranscripts_method}{use either "dbscan" or "delaunay method" to group transcripts in space (default = "dbscan")}

\item{distance_cutoff}{maximum molecule-to-molecule distance within same transcript group (default = "auto")}

\item{config_spatNW_transcript}{configuration list to create spatial network at transcript level, see manual for \code{createSpatialDelaunayNW_from_spatLocs} for more details, set to NULL to use default config}

\item{seed_transError}{seed for transcript error detection step, default = NULL to skip the seed}

\item{spatLocs_coln}{the column name of 1st, 2nd, optional 3rd spatial dimension of each transcript in transcript_df}
}
\value{
data frame for transcripts in \code{chosen_cells} only, containing information for transcript score classifications and spatial group assignments as well as new cell/group ID for downstream resegmentation.
}
\description{
modular wrapper to identify transcript groups of poor fit to current cell segments in space
}
