% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_transcripts.R
\name{scoreGenesInRef}
\alias{scoreGenesInRef}
\title{scoreGenesInRef}
\usage{
scoreGenesInRef(genes, ref_profiles)
}
\arguments{
\item{genes}{a vector of gene name to score}

\item{ref_profiles}{a gene X cell_type expression matrix for reference profiles}
}
\value{
loglik, a gene X cell_type matrix of loglik score for each gene
}
\description{
calculate log-likilhood score of each gene based on reference expression profiles
}
