% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegRefinement.R
\name{transDF_to_perCell_data}
\alias{transDF_to_perCell_data}
\title{transDF_to_perCell_data}
\usage{
transDF_to_perCell_data(
  transcript_df,
  transGene_coln = "target",
  cellID_coln = "updated_cellID",
  spatLocs_colns = c("x", "y", "z"),
  celltype_coln = "updated_celltype",
  return_cellMeta = FALSE
)
}
\arguments{
\item{transcript_df}{transcript data.frame}

\item{transGene_coln}{column name of target or gene name in \code{transcript_df}}

\item{cellID_coln}{column name of cell_ID in \code{transcript_df}}

\item{spatLocs_colns}{column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in \code{transcript_df}}

\item{celltype_coln}{column name for cell type in \code{transcript_df}}

\item{return_cellMeta}{flag to return per cell data.table for metadata like mean spatial coordinates and cell types}
}
\value{
a list
\describe{
\item{perCell_DT}{a per cell data.table with mean spatial coordinates, cell type, return when \code{return_cellMeta} = TRUE}
\item{perCell_expression}{a gene x cell count sparse matrix derived from transcript data.frame after resegmentation}
}
}
\description{
get per cell expression matrix and optional metadata like mean spatial coordinates and cell types from transcript data.frame
}
