% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resegment_decisions.R
\name{check_config_leiden}
\alias{check_config_leiden}
\title{check_config_leiden}
\usage{
check_config_leiden(config)
}
\arguments{
\item{config}{a list of config would be used to create spatial network using delaunay method only via Giotto::createSpatialNetwork}
}
\value{
the corrected config list
}
\description{
check config used for leiden clustering, assign default values if missing arguments
}
