% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastReseg-package.R
\docType{package}
\name{FastReseg-package}
\alias{FastReseg}
\alias{FastReseg-package}
\title{FastReseg: An R package for Detection and Correction of Cell Segmentation Error Based on Spatial Profile of Transcripts}
\description{
An R package for detection and correction of cell segmentation error based on spatial profile of transcripts
}
\author{
\strong{Maintainer}: Lidan Wu \email{lwu@nanostring.com}

Authors:
\itemize{
  \item Patrick Danaher \email{pdanaher@nanostring.com}
}

}
\keyword{internal}
