% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_transcripts.R
\name{getCellType_maxScore}
\alias{getCellType_maxScore}
\title{getCellType_maxScore}
\usage{
getCellType_maxScore(
  score_GeneMatrix,
  transcript_df,
  transID_coln = "transcript_id",
  transGene_coln = "target",
  cellID_coln = "cell_ID",
  return_transMatrix = FALSE
)
}
\arguments{
\item{score_GeneMatrix}{a gene x cell-type score matrix}

\item{transcript_df}{the data.frame of transcript_ID and cell_ID}

\item{transID_coln}{the column name of transcript_ID in transcript_df}

\item{transGene_coln}{the column name of target or gene name in transcript_df}

\item{cellID_coln}{the column name of cell_ID in transcript_df}

\item{return_transMatrix}{logic flag whether to return the transcript level of score matrix (default = FALSE)}
}
\value{
return a list
\enumerate{
\item{cellType_DF, a data.frame of cell_ID and cell_type}
\item{score_TransMatrix, a transcript x cell-type score matrix when return_transMatrix = TRUE}
}
}
\description{
get the cell type give maximum score
}
