% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_FastReseg.R
\name{myFun_fov_load}
\alias{myFun_fov_load}
\title{myFun_fov_load}
\usage{
myFun_fov_load(path_to_fov)
}
\description{
supporting function for \code{fastReseg_internalRef}, to load transcript data.frame of each FOV from file path
}
