% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_score_utilities.R
\name{scoreGenesInRef}
\alias{scoreGenesInRef}
\title{scoreGenesInRef}
\usage{
scoreGenesInRef(genes, ref_profiles, flag_center = TRUE)
}
\arguments{
\item{genes}{a vector of gene name to score}

\item{ref_profiles}{a gene X cell_type expression matrix for reference profiles}

\item{flag_center}{flag to center the score matrix per gene before turn, default = TRUE}
}
\value{
loglik, a gene X cell_type matrix of centered loglik score for each gene
}
\description{
calculate log-likilhood score of each gene based on reference expression profiles and return the centered score matrix
}
