% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_segments.R
\name{update_transDF_ResegActions}
\alias{update_transDF_ResegActions}
\title{update_transDF_ResegActions}
\usage{
update_transDF_ResegActions(
  transcript_df,
  reseg_full_converter,
  score_GeneMatrix = NULL,
  refProfiles = NULL,
  transGene_coln = "target",
  cellID_coln = "cell_ID",
  celltype_coln = "cell_type",
  spatLocs_colns = c("x", "y", "z"),
  return_perCellDF = TRUE,
  celltype_method = "LogLikeRatio"
)
}
\arguments{
\item{transcript_df}{the data.frame of transcript to be updated}

\item{reseg_full_converter}{a named converter to update the cell ID in \code{transcript_df}, cell_ID in name would be converted to cell_ID in value; discard cell_ID with value = NA}

\item{score_GeneMatrix}{the gene x cell-type matrix of log-like score of gene in each cell type, needed if using \code{LogLikeRatio} cell typing method (default  = NULL)}

\item{refProfiles}{A matrix of cluster profiles, genes X clusters, needed if using \code{NegBionomial} cell typing method (default  = NULL)}

\item{transGene_coln}{the column name of target or gene name in \code{transcript_df}}

\item{cellID_coln}{the column name of cell_ID in \code{transcript_df}}

\item{celltype_coln}{the column name of cell type in \code{transcript_df}}

\item{spatLocs_colns}{column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in transcript_df}

\item{return_perCellDF}{flag to return gene x cell count matrix and per cell DF with updated mean spatial coordinates and new cell type}

\item{celltype_method}{use either \code{LogLikeRatio} or \code{NegBinomial} method for quick cell typing and corresponding score_baseline calculation (default = LogLikeRatio)}
}
\value{
a list
\describe{
\item{updated_transDF}{the updated transcript_df with \code{updated_cellID} and \code{updated_celltype} column based on reseg_full_converter}
\item{perCell_DT}{a per cell data.table with mean sptial coordinates and new cell type when return_perCellDF = TRUE}
\item{perCell_expression}{a gene x cell count sparse matrix for updated transcript data.frame when return_perCellDF = TRUE}
}
}
\description{
Update transcript data.frame based on resegmentation action, calculate the new cell type and mean per cell spatial coordinates
}
\details{
Update transcript data.frame based on resegmentation action and get new cell type; when return_perCellDF = TRUE, return gene x cell count matrix and per cell data.frame with mean per cell spatial coordinates and new cell type.
}
